// BingAddressAutocomplete/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    value: ComponentFramework.PropertyTypes.StringProperty;
    bingapikey: ComponentFramework.PropertyTypes.StringProperty;
    street: ComponentFramework.PropertyTypes.StringProperty;
    city: ComponentFramework.PropertyTypes.StringProperty;
    county: ComponentFramework.PropertyTypes.StringProperty;
    state: ComponentFramework.PropertyTypes.StringProperty;
    zipcode: ComponentFramework.PropertyTypes.StringProperty;
    country: ComponentFramework.PropertyTypes.StringProperty;
    latitude: ComponentFramework.PropertyTypes.StringProperty;
    longitude: ComponentFramework.PropertyTypes.StringProperty;
    street1: ComponentFramework.PropertyTypes.StringProperty;
    city1: ComponentFramework.PropertyTypes.StringProperty;
    county1: ComponentFramework.PropertyTypes.StringProperty;
    state1: ComponentFramework.PropertyTypes.StringProperty;
    zipcode1: ComponentFramework.PropertyTypes.StringProperty;
    country1: ComponentFramework.PropertyTypes.StringProperty;
    latitude1: ComponentFramework.PropertyTypes.StringProperty;
    longitude1: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    value?: string;
    street?: string;
    city?: string;
    county?: string;
    state?: string;
    zipcode?: string;
    country?: string;
    latitude?: string;
    longitude?: string;
}